put hashing functions into own module
