+++ Cargo.toml
@@ -19,7 +19,6 @@ md5 = "0.7.0"
 rocket = { version = "0.5.0-rc.2", features = ["json"] }
 regex = "1"
 serde = { version = "1.0", features = ["derive"] }
-serde_repr = "0.1.10"
 serde_json = "1"
 getopts = "0.2"
 cached = "0.40.0"
